+++ tests/testthat/test_Read_and_Split_MutectVCFs.R
@@ -1,4 +1,4 @@
-context("Tests for internal functions reading and splitting mutect vcfs")
+context("Tests for internal functions reading and splitting vcfs")
 
 test_that("Test ReadMutectVCFs and SplitListOfMutectVCFs", {
   files <- list.files(path = "testdata/Mutect-GRCh37", full.names = TRUE)
@@ -8,4 +8,18 @@ test_that("Test ReadMutectVCFs and SplitListOfMutectVCFs", {
   split.vcfs <- SplitListOfMutectVCFs(vcfs)
   expect_null(split.vcfs$discarded.variants$`Mutect.GRCh37.s2`)
   
+  catalogs <- MutectVCFFilesToCatalog(files = files, ref.genome = "hg19",
+                                      region = "genome")
+})
+
+test_that("Test ReadStrelkaSBSVCFs and SplitListOfStrelkaSBSVCFs", {
+  files <- list.files(path = "testdata/Strelka-SBS-GRCh37/", full.names = TRUE)
+  vcfs <- ReadStrelkaSBSVCFs(files)
+  expect_equal(dim(vcfs[[1]]), c(798, 21))
+  
+  split.vcfs <- SplitListOfStrelkaSBSVCFs(vcfs)
+  expect_null(split.vcfs$discarded.variants$`Mutect.GRCh37.s2`)
+  
+  catalogs <- StrelkaSBSVCFFilesToCatalog(files = files, ref.genome = "hg19",
+                                          region = "genome")
 })